+++ glean-core/android/src/test/java/mozilla/telemetry/glean/private/AccumulationsBeforeGleanInitTest.kt
@@ -9,7 +9,7 @@ import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.work.testing.WorkManagerTestInitHelper
 import mozilla.telemetry.glean.Glean
-import mozilla.telemetry.glean.GleanMetrics.GleanBuildInfo
+import mozilla.telemetry.glean.GleanBuildInfo
 import mozilla.telemetry.glean.Dispatchers
 import mozilla.telemetry.glean.config.Configuration
 import org.junit.After